AG-14639 Move labelBoxOptionsDef into ag-charts-core to fix build
